apps-source-code/Python/requirements.txt
psycopg2-binary==2.9.1
Flask==2.2.2
Werkzeug==2.2.2